+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-21T13:39:04.321Z"
+    created: Optional[str] = "2026-07-21T15:01:26.715Z"
 
-    updated: Optional[str] = "2026-07-21T13:39:04.321Z"
+    updated: Optional[str] = "2026-07-21T15:01:26.715Z"
 
     updated_by_id: OptionalNullable[str] = UNSET