+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY2EENPH4TMCW66T79XF5WWZ"
+        "01KY2K4TGQECM9HXQRH78SN3BG"
     )
     r"""The unique identifier of the data source"""