+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY2EENPX3JGRBN3J9JB32E57"
+        "01KY2K4TH0G5X8XTNRC0P6KTC3"
     )
     r"""The unique identifier of the data source"""