+++ packages/orq-rc/RELEASES.md
@@ -9589,3 +9589,13 @@ Based on:
 - [python v4.12.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.47 - packages/orq-rc
+
+## 2026-07-21 15:04:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.48 - packages/orq-rc